Update:Design/db_ColumnType

+++ project/db/init.sql
@@ -18,7 +18,7 @@ create table books
     publisher text,
     pages integer,
     isbn text,
-    rating integer,
+    rating NUMERIC(3,1),
     description text,
     created_by integer,
     foreign key (created_by) references users(id) on delete set null
